delete all interpreter, resolver and type checking code
I started to get lost on how to implement type checking.
So I decided to start from scratch on that. First I need
to have the right ast nodes for each thing and complete
the parser. Maybe iterating at the same time on both the
parser and interpreter/resolver/typechecker is a mistake
and causes confusion. The syntax analysis should be
finished first before starting the semantic analysis.