restrict parsing of composite literals
The decision on which kind of type or identifier to
allow a composite literal is intricate. For instance,
it's not simple to allow a variable to have a
composite literal within an if statement's condition
expression: if num > max {...

In the example above, how can we decide if max is of
a type that allows a composite literal? Or the left
brace refers to the block of the if statement?